调整适配 sea-orm 并移除 postgresql 支持

+++ src/common/forms.rs
@@ -5,9 +5,9 @@ use validator::Validate;
 #[derive(Serialize, Deserialize, Validate)]
 pub struct ListQuery {
     #[validate(range(min = 1, message = "page must greater than 1"))]
-    pub page: Option<i32>,
+    pub page: Option<u64>,
     #[validate(range(min = 1, message = "page_size must greater than 1"))]
-    pub page_size: Option<i32>,
+    pub page_size: Option<u64>,
     #[validate(length(min = 1, max = 13, message = "order_by length must greater than 1"))]
     pub order_by: Option<String>,
 }
@@ -15,9 +15,9 @@ pub struct ListQuery {
 #[derive(Serialize, Deserialize, Validate)]
 pub struct ListQueryWithPrivate {
     #[validate(range(min = 1, message = "page must greater than 1"))]
-    pub page: Option<i32>,
+    pub page: Option<u64>,
     #[validate(range(min = 1, message = "page_size must greater than 1"))]
-    pub page_size: Option<i32>,
+    pub page_size: Option<u64>,
     #[validate(length(min = 1, max = 13, message = "order_by length must greater than 1"))]
     pub order_by: Option<String>,
     pub private: Option<bool>,
