+++ src/posts/forms.rs
@@ -5,9 +5,9 @@ use validator::Validate;
 #[derive(Serialize, Deserialize, Validate)]
 pub struct PostsQuery {
     #[validate(range(min = 1, message = "page must greater than 1"))]
-    pub page: Option<i32>,
+    pub page: Option<u64>,
     #[validate(range(min = 1, message = "page_size must greater than 1"))]
-    pub page_size: Option<i32>,
+    pub page_size: Option<u64>,
     #[validate(length(min = 1, max = 13, message = "order_by length must greater than 1"))]
     pub order_by: Option<String>,
     pub private: Option<bool>,
@@ -20,7 +20,7 @@ pub struct PostCreate {
     pub title: String,
     #[validate(length(min = 1, max = 150, message = "slug length must greater than 1"))]
     pub slug: String,
-    pub created: i32,
+    pub created: u32,
     pub text: String,
     #[validate(length(min = 1, max = 16, message = "status length must greater than 1"))]
     pub status: String,
